future: store result in OnceLock
The result is going to be initialized only once, thus we do not need to
store it behind a mutex. We can use OnceLock instead. Thanks to that,
we can remove the unsafe logic which extends the lifetime of `coordinator`
reference in cass_future_coordinator. We now guarantee that
the result will be immutable once future is resolved - the guarantee
is provided on the type-level.