future: implement cass_future_coordinator
There is one tricky part that we need to handle: the coordinator is held
by CassResult, which is stored under Mutex in CassFuture. In result, the
reference we obtain in `CassFuture::with_waited_result` closure has a lifetime
of the mutex guard (temporary lifetime during the function call). We need to
extend the lifetime of returned coordinator - further reasoning is explained in
the comment in code.