integration: increase timestamp tolerance in Timestamps test
The Timestamps test in the integration suite was failing due
to the timestamp difference being bigger than the tolerated threshold.
Git blame reveals that the threshold was alread increased from 100ms
to 150ms by Gor in 2022. I assume 500 ms will be enough for now.

A question aside is whether this test comparison gives us any benefit
at all. It seems to be a flaky test that is not very useful.
Perhaps we should consider just removing this flaky timing-based check.