testing: implement a retry policy that ignores all errors
IgnoringRetryPolicy is a retry policy that always ignores all errors.
It's implemented as another variant of CassRetryPolicy, and it is going
to be used in integration tests. There is one test in particular,
`ExecutionProfileTest.Integration_Cassandra_RetryPolicy`, which requires
this retry policy to be able to run. In order to enable that test and
not change the retry policy it uses, we need to implement it.

The next commits will implement the bridge between Rust and C++
integration testing module, so that the test can use this retry policy.