codewide: privatise unnecessarily public items
This is a big commit that privatises a lot of items which have been
`pub` for no good reason. The goal is to make the public API
of the wrapper clear and minimal, and to prevent silencing warnings
about unused items (because `pub` items are always considered used).

The warnings this commit triggers are fixed in the next few commits.