codewide: privatise unnecessarily public items
This is a big commit that privatises a lot of items which have been
`pub` for no good reason. The goal is to make the public API
of the wrapper clear and minimal, and to prevent silencing warnings
about unused items (because `pub` items are always considered used).

I've kept the `argconv` module public, together with its traits and
types, because it's used in doctests. Privatising it would make them
stop compiling. Also, those type and traits are actually part of the
API, in a sense that they are FFI-facing. So let them stay public.