lbp: `used_hosts_per_remote_dc=0` disables DC failover
The parameter `used_hosts_per_remote_dc` of the DC-aware LBP has been
long deprecated. However, it's been still supported in the cpp-driver.
Due to Rust driver API's, it'd be very hard to support it correctly
in the case of `>0`, because of reasons explained in
<https://github.com/scylladb/cpp-rust-driver/issues/315>.
Therefore, the partial support approach is taken:
- case `=0` -> employ `HostFilter`, disable DC failover and we're done.
  Our semantics here are in line with the CPP Driver.
- case `>0` -> Treat this as `+inf` case - don't limit connections nor
  requests to remote nodes. Emit a warning that the semantics are
  different than in CPP Driver.

This commit only configures the LBP to set `permit_dc_failover` to false
when `used_hosts_per_remote_dc` is `0`, and to true when it's greater
than `0`.

The next commit will implement the connection management based on this
parameter: If a DC is remote to **all** LBPs **and** those LBPs disallow
DC failover, then the driver will not connect to any hosts in that DC.
This will finish proper support for the `used_hosts_per_remote_dc=0`
case.