lbp: `used_hosts_per_remote_dc=0` prevents opening connections
The parameter `used_hosts_per_remote_dc` of the DC-aware LBP has been
long deprecated. However, it's been still supported in the cpp-driver.
Due to Rust driver API's, it'd be very hard to support it correctly
in the case of `>0`, because of reasons explained in
<https://github.com/scylladb/cpp-rust-driver/issues/315>.
Therefore, the partial support approach is taken:
- case `=0` -> employ `HostFilter`, disable DC failover and we're done.
  Our semantics here are in line with the CPP Driver.
- case `>0` -> Treat this as `+inf` case - don't limit connections nor
  requests to remote nodes. Emit a warning that the semantics are
  different than in CPP Driver.

This commit introduces specific connection management for the case
when `used_hosts_per_remote_dc=0`: if a DC is remote to **all** LBPs
**and** those LBPs disallow DC failover, then the driver will not
connect to any hosts in that DC.

Implementation is done using the existing `Filtering{Config,Info}`
machinery. LBPs declare whether they restrict DCs to the specific local
DC, or allow all DCs - `DcRestriction` enum serves this purpose.
Then, `CassHostFilter` is built from the composition of DC restrictions,
by building up an `AllowedDcs` enum: if any LBP allows all DCs,
then `AllowedDcs::All` is used, otherwise, we build a whitelist of
allowed DCs is constructed (`AllowedDcs::Whitelist(Vec<String>)`).
Finally, if `AllowedDcs` is `Whitelist`, then `CassHostFilter` filters
out hosts whose DCs are not in that whitelist.

This finishes proper support for the `used_hosts_per_remote_dc=0` case.