Partly revert "tests: adjust DCExecutionProfileTests constructor"
This partly reverts commit 6b44c2cb5d2ab82e1674d4e83c7fbd2266c263bf.
This is because now we support `used_hosts_per_remote_dc` correctly,
which means we forbid DC failover in the case of
`used_hosts_per_remote_dc = 0`. This makes the test fail, because it
originally had that parameter set to 1. The reverted commit changed it
to `0`, because `0` was the only accepted parameter at that time.