IT: enable DCExecutionProfileTest_DCAware test
The test DCExecutionProfileTest.DCAware was disabled because it expected
deterministic behaviour of the token-unaware round robin load balancing.
As the Rust Driver round-robins nodes with a random shift, the test
would fail. I reworked it to check that the node in the other DC is not
used, instead of expecting a specific node in the local DC to be used.

The test is now enabled in the Makefile.