statement: implement recording attempted hosts
This is the first of the two commits that implement and expose recording
hosts that were attempted during execution of a statement.
This is useful for testing purposes, allowing us to verify which hosts
were attempted during a request.
In particular, some existing integration tests rely on this feature.
Implementing this will allow us to run those tests, as well as write
new tests with more assertions regarding routing.

Implementation involves:
- Implementing `RecordingHistoryListener` that records attempted hosts;
- Adding exposed `testing_statement_set_recording_history_listener` to
  enable recording in the statement's history listener;
- Modifying the session execution to set the history listener if
  recording is enabled;
- Adding a method to retrieve the recorded hosts from the listener
  (which will be used in the next commit).