future: expose recorded attempted hosts
This commit exposes the `attempted_hosts` method on `CassFuture`,
allowing users to retrieve the hosts that were attempted during
the execution of a future. This can be used in proxy tests in the
wrapper itself, but the main use case is in the C++ integration
tests. The necessary bridging code is implemented in the next commit.