IT: implement `testing_future_get_attempted_hosts`
This commit implements the function `testing_future_get_attempted_hosts`
in Rust, as well as includes a test for the new function. It's going to
be used in the C++ integration tests.

I had tough time figuring out how to return a list of attempted hosts'
addresses to C code, so I decided to return a concatenated string of
them in a stringified form, delimited by '\n'. The overhead this
introduces is not important at all, as these are just integration tests.

The caller is responsible for freeing the memory. They must do so by
calling `testing_free_cstring`, because the string is allocated in Rust.

The new unit test, unlike all unit tests so far, is put behind
the #[cfg(cpp_integration_testing)] attribute. In order to run it,
we must to supply the `cpp_integration_testing` flag to RUSTFLAGS.
So far, RUSTFLAGS were configured in .cargo/config.toml, in `[build]`.
As this method of setting RUSTFLAGS does not allow setting different
flags for test target and for the rest of the build, I decided to
set the necessary RUSTFLAGS for tests in the Makefile. Keep in mind
that this will require us to keep the flags there and in the
`.cargo/config.toml` in sync, unfortunately.