CI: don't test with Scylla 5.4.8
The Scylla 5.4.8 was introduced to the CI 9 months ago, in
acf9d3d23b794d19b9dd370dff0b07328ed58e95. The rationale was to test on
the same version as the cpp-driver CI. There's no use in testing neither
cpp-driver nor this driver with such old version anymore.

As some tests fail due to ccm errors on that version, we remove it from
the CI.