IT: include error message in the exception string
When an exception is thrown in integration tests, the stringified error
code would be included in the exception message, but not the error
message. In order to make debugging easier, the error message is now
included in the exception message as well. This exhibits the underlying
Rust Driver's nested errors' messages.