session: allow case-sensitive keyspace names
The CPP Driver allows keyspace names to be case-sensitive if they are
enclosed in quotes. CPP-Rust Driver, when encountered a quoted keyspace,
would reject it with an error, because it verifies that a keyspace name
only contains alphanumeric characters and underscores.
Rust Driver does allow case-sensitive keyspace names, but this setting
must be explicitly set by a boolean flag, and the name itself must not
be quoted. This commit adds a check for quoted keyspace names, and if
they are found, it removes the quotes and sets the case-sensitivity flag
accordingly.
This allows us to enable another integration test:
`UseKeyspaceCaseSensitiveTests.ConnectWithKeyspace`.