IT: negated qualification of MetricsTests
As we now run more tests from the MetricsTests suite than not,
it makes sense to negate the qualification of the tests, so that we
say that we run the whole suite _except_ two tests.

This way of specifying disabled tests explicitly makes them more
discoverable for further development, so that we can easier see what's
not yet supported and/or tested.