TMP: tune AsyncTests.Close to fail with current impl
The current implementation of the driver is flawed. In order
to empirically prove this, I tune the AsyncTests.Close test
to fail with the current implementation. The tuning involves
changing the number of concurrent requests to 5 and adding
a sleep time of 100 milliseconds between each request.
Moreover, the runtime is changed to use a multi-threaded
runtime with 2 worker threads to allow for concurrent execution
with interleaving that promotes the issue.

This commit is intended to demonstrate the issue with the current
implementation of the driver, specifically in how it handles
session closure concurrent to running requests. The test should fail
because some requests will only try to read-lock the session
after the session has been closed, leading to failure of those requests.

The following commits will address the issue by read-locking the session
synchronously upon scheduling a request (via either of
`cass_session_execute[_batch]`).

This commit shall be removed before merging this PR into master.