session: don't lock blockingly
There is number of places in the code where we use `blocking_read()`
to access the session data. This is suboptimal because it blocks the
current thread, which can lead to deadlocks (on current_thread
executor). To accomodate for this case, we switch to using
`RUNTIME.block_on(lock.read())` to allow the async runtime to work
on queued tasks while waiting for the lock to be released.