session: lock synchronously to prevent logical races
AsyncTests.Close test specifies that the session only be closed after
all running requests are finished. This is not necessarily the case
in the current implementation, because the session lock is taken for
reading by the request futures only in the async block. Races may thus
occur if the session is closed while the request is being executed
and it has not yet acquired the lock, leading in request failures.

To be in line with the CPP Driver, we take the session lock
synchronously, before returning from `cass_session_execute[_batch]`
functions. This way we ensure that the session is not closed
while there are any requests running, and that the session is closed
only after all requests are finished.

Care is taken to ensure that the current_thread runtime does not deadlock
when waiting for the lock, by executing remaining futures while waiting
for the lock to be released (by calling `RUNTIME.block_on`).