examples tuple,udt: Expect TEXT or VARCHAR
The examples/tuple/tuple.c and examples/udt/udt.c tests expect VARCHAR
to be the type of the text field in the tuple and UDT respectively.
The problem is, the Rust driver merges VARCHAR and TEXT into a single
enum variant: ColumnType::Text, mimicking the fact that both types are
exactly the same in ScyllaDB and Cassandra. As a result, Rust driver
reports the type as TEXT, while the examples expect it to be VARCHAR.
This results in the following output of the tuple example:

`id 98798b00-5340-11f0-878a-2faf302264f1 <invalid> 1292624629956446976`

This commit updates the examples to expect TEXT as well, so that their
output becomes correct. Tuple example output:

`id a9457840-5340-11f0-a7db-a9db64eeb5c9
"a9457840-5340-11f0-a7db-a9db64eeb5c9" 1292624630238246976`

Similarly, the UDT example is fixed.