examples: add `drop_examples_keyspace` example
The example will be used by the example runner in Makefile to drop
the `examples` keyspace between runs of consecutive examples.
This is because the examples often create the `example` keyspace,
don't drop it, and at the same time expect that the keyspace does not
exist when they run. To avoid this, we add a new example that drops the
`examples` keyspace, and the example runner will run it before running
the next example.
Alternatively, we could use cqlsh, but this would require adding it
as a new dependency, which is not desirable.