exec_profile: fix serial consistency handling
Similarly to previous commits, this change ensures that the
serial consistency is set to `None` when the value is
`CASS_CONSISTENCY_ANY`. This aligns with the behavior of the CPP Driver.

Note that I previously mistreated the `CASS_CONSISTENCY_UNKNOWN` value
as an "any" value. This is incorrect: `CASS_CONSISTENCY_UNKNOWN` denotes
"ignore me", which means that if for example a statement has
`CASS_CONSISTENCY_UNKNOWN` set, then the execution profile should be
considered instead. To compare, `CASS_CONSISTENCY_ANY` set on a
statement overrides the execution profile.

Actually, if a CPP execution profile specifies
`CASS_CONSISTENCY_UNKNOWN` for its setting, then this setting is taken
from the cluster = from the default execution profile (which forbids
`CASS_CONSISTENCY_UNKNOWN`, which is ensured in the next commit).
I'm going to introduce the described mechanism in a follow-up PR.
For now, `CASS_CONSISTENCY_UNKNOWN` passed to
`cass_execution_profile_set_serial_consistency` results in an error
`CASS_ERROR_LIB_BAD_PARAMS` returned.

The change is done by leveraging the new `MaybeUnsetConfig` type
to represent a consistency value that can be either set or unset.