prepared: Fix consistency bug
The fixed bug is quite severe: any prepared statement would have its
consistency set to `One` by default. This is surprising, because `One`
hasn't been the default consistency for prepared (or any other kind of)
statements in CPP Driver since 2015 (as per the changelog)...

Anyway, the CPP Driver's behaviour upon preparing is to set consistency
in prepared statements to `UNKNOWN`, which means "use the consistency
configured in the exec profile/cluster". This is the behaviour of Rust
Driver, too: `PreparedStatement` has no consistency set by default.

Thus, the fix is quite simple: we just remove the line that sets
consistency to `One` in the `cass_session_prepare_n` function.