cluster: use MonotonicTimestampGenerator by default
This is in line with the CPP Driver.

Also, fixed the leftover old comment about the default value of the
timestamp generator, which has not correct anymore for some years.
When the DataStax changed the default value of the timestamp generator
from `ServerTimestampGenerator` to `MonotonicTimestampGenerator`, they
forgot to update the comment in `cassandra.h`.