cluster: set default policies explicitly
The default execution profile in the Scylla Rust wrapper
has been setting default retry policy and speculative execution policy
correctly (in line with the CPP Driver). However, it's a better
defensive pattern to set these explicitly in the default execution
profile builder, so that any future changes to the default execution
profile in the Rust Driver do not break the Scylla Rust wrapper.