cluster: describe LBP divergence from CPP Driver
CPP Driver uses DCAware load balancing policy by default.
The tricky part is that it initializes the local DC to the first node
the client connects to. This is tricky, a possible footgun, and hard to
be done with the current Rust driver implementation, which is why we use
RoundRobin by default (we turn off DC awareness by default, requiring
the user to provide the local DC explicitly if they want to use it).

The comments describing this behaviour are added to the `cluster.rs`
and `load_balancing.rs` files for better discoverability and
understanding of this divergence.