exec_profile: use cluster defaults for unset settings
This commit modifies the `CassExecProfile` to use the defaults from
the cluster's default profile for settings that are not explicitly
set in the execution profile. This is consistent with the behavior
of the CPP driver.
Note that the consistency is handled separately, as it has a separate
logic in CPP Driver (whether it makes sense or not is another topic
for dicussion): if consistency is not set in the execution profile,
it is set to the hardcoded default consistency, not the one from
the cluster's default profile.