iterator: hide the Iterator enum from the public API
The variants of the `CassIterator` should not be exposed to the public
API, as they are an implementation detail of the Scylla Rust wrapper.
To achieve this, the `CassIterator` enum is now renamed to
`CassIteratorInner` and wrapped in a new `CassIterator` struct,
whose only purpose is to provide a pub-opaque wrapper around the inner
enum. This allows us to keep the public API clean.

Without this change, the compiler complained when I pubified
the `CassIterator` enum (for use in Rust integration tests) without
pubifying its variants' enums.