introduce config_value module
The module `config_value` is introduced to handle configuration values
that can be either set or unset. It features a trait
`MaybeUnsetConfigValue` and an enum `MaybeUnsetConfig` that work in
tandem to represent configuration values that may or may not be set.

`MaybeUnsetConfigValue` is a Rust type that can be either set or unset,
together with a constant defining the special "unset" C value and with
methods to obtain the Rust type from a C value.

For now, implementations of `MaybeUnsetConfigValue` are provided for
Consistency and Option<SerialConsistency>, which are used in the
next commits to handle (serial) consistency in a more robust way.
In the future, more implementations can be added to this module as
needed, with request timeout duration and retry policy being
my first candidates.

Intended usage will be shown soon, when those types are used in the
configuration code.