statement,batch: refactor UNKNOWN consistency handling
CPP Driver allows setting `CASS_CONSISTENCY_UNKNOWN` on statement
or batch to indicate that consistency is not set, which means
"use whatever the execution profile/cluster specifies". Let's
ensure consistent behaviour in CPP-Rust Driver.

Unfortunately, the Rust Driver does not support "unsetting" consistency
from a statement at the moment. For now, we will throw an error upon
`CASS_CONSISTENCY_UNKNOWN` provided, in order to warn the user about
this limitation.

The change is done by leveraging the new `MaybeUnsetConfig` type
to represent a consistency value that can be either set or unset.