cluster: refactor UNKNOWN consistency handling
CPP Driver forbids setting `CASS_CONSISTENCY_UNKNOWN` on cluster
to indicate that consistency is not set, because cluster must specify
some default consistency to be used if there's no execution
profile-level nor statement-level consistency specified. This behaviour
has already been replicated in the Rust wrapper.

This commit merely leverages the new `MaybeUnsetConfig` type
to represent a consistency value that can be either set or unset,
in order to unify the handling of unset consistency values and
prevent bugs.