exec_profile: use MaybeUnsetConfig for request timeout
This refactors the function `cass_execution_profile_set_request_timeout`
to use the `MaybeUnsetConfig` type for the timeout argument. This is
another step towards unifying the way maybe-unset configuration options
are handled in the Scylla Rust wrapper.

Additionally, this change fixes a bug that the request timeout was
not being turned off when the timeout was set to 0. The bug would cause
immediate timeouts in the driver, as the request timeout was set to 0.

The unit test for execution profiles has been updated to check
correctness after this change.