statement: fix request timeout handling
Request timeout handling in `cass_statement_set_request_timeout` was
broken: it was implemented by hand, which meant that it did not
cooperate with execution profiles. In particular, if someone set
a longer timeout on a statement than the one set in the execution
profile, then the execution profile's timeout was still used.

This commit fixes this by using the dedicated Rust Driver Statement's
API for setting the request timeout, which correctly cooperates with
execution profiles (including the cluster's default request timeout).