batch: partially fix request timeout handling
Request timeout handling in `cass_batch_set_request_timeout` was
broken:
1. It was implemented by hand, which meant that it did not
cooperate with execution profiles. In particular, if someone set
a longer timeout on a statement than the one set in the execution
profile, then the execution profile's timeout was still used.
2. It did not handle the `0` value correctly - it was interpreted
as a request timeout of 0 milliseconds, while it should be interpreted
as "no timeout".

This commit fixes the second problem.
The first problem must wait for the Rust Driver to expose a way to set
the request timeout on a batch.