support building examples straight from CMake
After the previous commit, the examples could be built using `make`.
This commit adds support for building examples straight from CMake, by
running `cmake -DCASS_BUILD_EXAMPLES=on -DCMAKE_BUILD_TYPE=Release ..`
in the `build` directory. As here it's `CMake` that parses Makefile,
not `make` that runs `CMake`, we need to extract the list of examples
to run from the Makefile. This is done by running `make -p -n` and
extracting the value of `SCYLLA_EXAMPLES_TO_RUN` variable.