Makefile: support running examples
The `run-examples-scylla` target is added to the Makefile to run
examples. The examples are built and run against a ScyllaDB instance
running in a Docker container. The `SCYLLA_EXAMPLES_TO_RUN` variable
is used to specify which examples to build and run.

As some examples create the `examples` keyspace and some assume it is not
present, a special example `drop_examples_keyspace` is added to drop the
`examples` keyspace before running any example.
This lets us avoid adding a dependency on `cqlsh` to the system.