future: return immediately if result is available
Fixes the most problematic instance of #329, when a panic is triggered
if `cass_future_error_code` (or any other function that await a future)
is called from a callback.

The direct cause of the panic is that a tokio executor thread calls
Runtime::block_on, which is illegal from the tokio context.
The cause is removed by checking if the result is already available,
and, if so, returning immediately, without trying to block on the
future task's `JoinHandle`.
Although this solves only a subset of the cases, it is a good start.
Also, the `callbacks` example now works!

Before this PR, there was an invariant that `join_handle` was always
awaited and set to None before result from the future could be accessed.
It was done in the `with_waited_result_timed` and `with_waited_result`
functions.
Now this invariant is no longer true. When future is created, it is
spawned on the runtime. `with_waited_result_timed` and
`with_waited_result` now have a happy path, that just checks existence
of the result. If result was set before call to those functions, is is
possible for `join_handle` waiting to be completely avoided, resulting
in situation with future result being used while `join_handle` is
`Some`.