session: migrate from RwLock to ArcSwapOption
This change gets rid of locking in the session, which is good both for
performance and for avoiding blocking the reactor thread, which has
shown to cause panics.

As the `CassConnectedSession` is now always stored behind an `Arc`,
it becomes shareable, and thus can be stored in request futures
independently of the `CassSession` itself. This lays the groundwork
for the robust session close handling, which will be implemented
in the next commits.