Makefile: run clippy over the IT code
Some code related to integration testing is conditionally compiled
based on the `cpp_integration_testing` flag, which is not set by
default. This made some lints not run on that code, which is
undesirable. This commit adds the `cpp_integration_testing` flag
to RUSTFLAGS when running clippy, so that all lints are run on
that code as well.

Keep in mind that `cargo check` is still run without the
`cpp_integration_testing` flag. This is done on purpose, to have one
basic check for the code with the `cpp_integration_testing` flag off
(to make sure it passes them correctly), and one extended (clippy) check
for the code with the `cpp_integration_testing` flag on.