session: pend/unpend requests when appropriate
This commit introduces the first part of the fix for the session close
issue. It modifies the session handling to set requests as pending
once they are started, and to unpend them once they are completed.

Care is taken to ensure that once a request has been marked as pending,
it will be eventually unpended. This is done by leveraging the type
system. `PendingRequestProof` is a marker type that has two goals:
1. It ensures that if a request is marked as pending, it will
   eventually be unpended. This is achieved by panicking upon drop
   unless the proof is disarmed by calling `unpend_request`.
   Note that this is not a strong guarantee, as it relies on the
   destructors being called, which is a not a Rust guarantee
   (think `mem::forget`, cyclic references, etc.).
2. It ensures that the request can be unpended only if it has been
   previously marked as pending. This is achieved by making
   `unpend_request` function take a `PendingRequestProof` as an
   argument. As the proof is only constructible from withing the
   defining module, this prevents misuse.

The `PendingRequestProof` is intended to be operated by the higher-level
abstraction, namely `CassConnectedSession::make_request_future`.
It takes care of requiring the request to have been pended synchronously
prior to the future creation, and it also takes care of unpending
the request once the future is resolved.
It has slightly inconvienient API, as it requires the wrapped future
to return a pair `(Arc<CassConnectedSession>, CassFutureResult)` instead
of the usual `CassFutureResult`. This is because due to lack of proper
support of `AsyncFn` traits (RTN not yet there, so I could not require
the `Future` returned by the `AsyncFn` closure to be `Send`), I failed
to avoid cloning the `Arc<CassConnectedSession>` in a more convenient
way. I hope to improve this in the future, once `AsyncFn` traits
get proper support in Rust.