config_value: associated type -> type parameter
As @lorak-mmk smartly pointed out, the `MaybeUnsetConfigValue` trait
should use a type parameter for the C value instead of an associated
type. This allows us to implement conversion from multiple C value types
to the same Rust type. Another benefit is that it allows us to
implement `MaybeUnsetConfigValue` for types that are not
`'static`, which will soon be useful for `Option<&RetryPolicy>`.

This commit refactors the `MaybeUnsetConfigValue` trait, as well as
adds a `CValue` type parameter to the `MaybeUnsetConfig` enum.
Without that, Rust refused `CValue` blanket type parameters in the
`impl` block for `MaybeUnsetConfig`, because it was not constrained
in any way.
Fortunately, this type parameter will be easily inferred in most cases,
so this won't bring any fuss.