session: fix race with `cass_session_close`
@lorak-mmk noticed the following bug in the closing protocol:
Upon concurrent execution of a request function, e.g.,
`cass_session_execute`, and `cass_session_close`, the following race may
occur:
1. Thread 1 calls `cass_session_execute`, which loads the
`CassConnectedSession`.
2. Thread 2 calls `cass_session_close`, which closes the session
   by swapping the `connected` field to `None`.
3. Thread 2 sets the `Notify`.
4. Thread 2 checks the number of pending requests, which may be 0
   because Thread 1 has not yet pended the request. It sees 0,
   concludes that all requests have finished, and resolves the close
   future.
5. Thread 1 calls `pend_request`, which increments the number of
   pending requests, but the session is already closed.
   Then it proceeds to make the request, which will succeed, even though
   the session is closed and the caller of `cass_session_close` has
   already awaited the future, so they believe that all requests have
   completed.

We must not start the request if the session is now closed.
To ensure this, we need to make sure that once we pended the request,
the session is still connected, i.e., the `Arc<CassConnectedSession>`
is still there and, to prevent ABA problem, that it is the same as
the one we loaded before pending the request. We verify that by checking
pointer equality of the `CassConnectedSession` Arc.
A little paranoia never hurts, especially in the FFI code ;)
This is what Rust devs are known for, right?

Credits to @lorak-mmk for noticing this bug.