future: safely disallow null callback pointers
The CPP Driver unsafely dereferences the callback pointer even if it is
null. Our Rust wrapper `unwrap()`s the Option with the callback, so it
panics in such case. This commit changes the code to return an error
upon passing a null callback pointer to `cass_future_set_callback`.
Then, the callback is always non-null if set in the driver, so no panic
is possible.

This should be considered an *improvement* compared to the CPP Driver.