gladly unset paging state if null ptr passed
The code of `cass_statement_set_paging_state_token` was unsetting the
paging state on a statement if a null paging state pointer was passed,
but it also returned an error in this case.
This is inconsistent with the behavior of other setters, which either
preserve the existing state and return an error if a null pointer is
passed, or unset the state and return CASS_OK.
For consistency and ease of use, the code is changed to unset paging
state when null is provided.
As a bonus, to mimic CPP Driver behaviour, empty strings now also
unset the paging state.

Note: CPP Driver does not check the passed pointer for null at all.