introduce CollectionType enum for type safety
The driver was using the C enum `CassCollectionType` directly in its
`CassCollection` struct, which led to necessity of handling the "else"
case (values other than the explicitly listed variants) in many places.

To bring type safety to the driver internals, this commit introduces a
new Rust `CollectionType` enum, which is used internally instead of the
C enum.