ci: remove old OSS Scyllas from workflows
The workflow that builds, lints, and tests the driver ran the tests
against four Scylla versions:
- ENTERPRISE-RELEASE
- ENTERPRISE-PRIOR-RELEASE
- OSS-RELEASE
- OSS-PRIOR-RELEASE

The OSS versions are no longer supported, so we remove them from the
workflow. It makes sense to keep the last two Scylla versions in the CI.
Other versions will be tested by the matrix tests.