cass_error: return INVALID_DATA where appropriate
`CASS_ERROR_LIB_INVALID_DATA` is returned in CPP Driver in cases where
the CQL values to be serialized/deserialized are not valid.
OTOH, `CASS_ERROR_LIB_UNEXPECTED_RESPONSE` is used when the message
is valid in terms of binary protocol, but the driver does not expect it
as a response in the current context.
I decided that errors in frame deserialization warrant
`INVALID_DATA` more then `UNEXPECTED_RESPONSE`.