cass_error: stop abusing server protocol error
`CassError` is split into categories, by the issuer of the error:
- `CASS_ERROR_LIB_*` for errors from the driver,
- `CASS_ERROR_SERVER_*` for errors from the server,
- `CASS_ERROR_SSL_*` for errors from the SSL library.

The `CASS_ERROR_SERVER_PROTOCOL_ERROR` was used for errors that are not
from the server, but from the driver, which is not correct. Among
other things, it confuses users of the library, who expect SERVER errors
to have come from the server, but they are actually from the driver.
To fix this, such usages of the error were replaced with
`CASS_ERROR_LIB_INVALID_STATE`.